Move template tags from cms app to readux app
