+++ apps/templates/snippets/volume_result.html
@@ -1,4 +1,4 @@
-{% load readux_templatetags %}
+{% load readux_extras %}
 
 <li class="uk-width-1-1@m uk-margin-small">
     <h4>
